Jukebox widget and hotkeys
This PR adds a new button in the jukebox tab that opens a small modal
with which you can control the jukebox playback, playlists and master
volume. The changes also introduce keyboard shortcuts that allow the
user to play specific playlists, skip to the next shuffled track in the
current playlist and stop all playback.